dom/security/test/mixedcontentblocker/test_bug1474842.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Bug 1474842: mixed content blocker bypass for localhost pages in iframe</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script>
    ok(window.isSecureContext, "top level should be a secure context");
    SimpleTest.waitForExplicitFinish(); 
    SpecialPowers.pushPrefEnv({set: [
      ["security.mixed_content.block_active_content", true],
      ["network.proxy.allow_hijacking_localhost", true],
      ["network.proxy.testing_localhost_is_secure_when_hijacked", true],
    ]}, () => {
      info("security.mixed_content.upgrade_display_content " + SpecialPowers.getBoolPref("security.mixed_content.upgrade_display_content"));
      info("security.mixed_content.block_display_content " + SpecialPowers.getBoolPref("security.mixed_content.block_display_content"));
      info("security.mixed_content.block_active_content " + SpecialPowers.getBoolPref("security.mixed_content.block_active_content"));
      info("security.mixed_content.block_object_subrequest " + SpecialPowers.getBoolPref("security.mixed_content.block_object_subrequest"));
      window.addEventListener("message", function listener(event) {
        if(!event.data.status) {
          return;
        }
        ok(event.data.status.startsWith("blocked"), "http script loaded from http://localhost should be blocked");
        window.removeEventListener("message", listener);
        SimpleTest.finish();
      });
      let f = document.createElement("iframe");
      f.src = "http://localhost/tests/dom/security/test/mixedcontentblocker/file_bug1474842.html";
      document.body.appendChild(f);
    }); 
  </script>
</head>

<body></body>
</html>
